refactor: better organize judging admin pages
