+++ src/pages/admin/judging/index.tsx
@@ -1,0 +1,233 @@
+import React from "react";
+import Head from "next/head";
+import Drawer from "../../../components/Drawer";
+import { CSVUploader } from "../../../components/CSVUploader";
+import { trpc } from "../../../utils/trpc";
+import { GetServerSidePropsContext, GetServerSidePropsResult } from "next";
+import { Role } from "@prisma/client";
+import { rbac } from "../../../components/RBACWrapper";
+import { getServerAuthSession } from "../../../server/common/get-server-auth-session";
+
+const JudgingPage: React.FC = () => {
+  const [status, setStatus] = React.useState<
+    "idle" | "loading" | "success" | "error"
+  >("idle");
+  const [errorMessage, setErrorMessage] = React.useState<string>("");
+  const [projectsPerTable, setProjectsPerTable] = React.useState<number>(10);
+  const [startTime, setStartTime] = React.useState<string>(
+    new Date().toISOString().slice(0, 16)
+  );
+  const [durationMinutes, setDurationMinutes] = React.useState(10);
+  const [judgingDuration, setJudgingDuration] = React.useState<string>("");
+  const [endTime, setEndTime] = React.useState<string>("");
+  const [numTables, setNumTables] = React.useState<number | null>(null);
+
+  const createTables = trpc.project.createTables.useMutation({
+    onSuccess: () => {
+      createTimeSlots.mutate({
+        slotDurationMinutes: durationMinutes,
+        startTime: new Date(startTime).toISOString(),
+      });
+    },
+    onError: (error) => {
+      setStatus("error");
+      setErrorMessage(error.message);
+      setTimeout(() => setStatus("idle"), 3000);
+    },
+  });
+
+  const createTimeSlots = trpc.timeSlot.createTimeSlots.useMutation({
+    onSuccess: (data) => {
+      setStatus("success");
+      if (data.endTime) {
+        const start = new Date(startTime);
+        const end = new Date(data.endTime);
+        const durationMs = end.getTime() - start.getTime();
+        const hours = Math.floor(durationMs / (1000 * 60 * 60));
+        const minutes = Math.floor(
+          (durationMs % (1000 * 60 * 60)) / (1000 * 60)
+        );
+        setJudgingDuration(`${hours} hours and ${minutes} minutes`);
+
+        setEndTime(
+          end.toLocaleTimeString("en-US", {
+            hour: "2-digit",
+            minute: "2-digit",
+            hour12: false,
+          })
+        );
+      }
+      if (data.numTables) {
+        setNumTables(data.numTables);
+      }
+    },
+    onError: (error) => {
+      setStatus("error");
+      setErrorMessage(error.message);
+      setTimeout(() => setStatus("idle"), 3000);
+    },
+  });
+
+  return (
+    <>
+      <Head>
+        <title>Judging Admin - DeltaHacks</title>
+      </Head>
+      <Drawer>
+        <main className="px-7 py-16 sm:px-14 md:w-10/12 lg:pl-20 2xl:w-8/12 2xl:pt-20 mx-auto max-w-4xl">
+          <h1 className="mb-12 text-2xl font-semibold leading-tight text-black dark:text-white sm:text-3xl lg:text-5xl 2xl:text-6xl text-center">
+            Judging Administration
+          </h1>
+
+          {/* Navigation Buttons */}
+          <div className="flex justify-center gap-4 mb-12">
+            <a href="/admin/judging/rubric" className="btn btn-accent">
+              Manage Judging Rubric
+            </a>
+            <a href="/admin/judging/leaderboard" className="btn btn-secondary">
+              View Leaderboard
+            </a>
+          </div>
+
+          <div className="grid gap-8 md:grid-cols-2">
+            {/* CSV Upload Section */}
+            <section className="card bg-base-100 shadow-lg">
+              <div className="card-body">
+                <h2 className="card-title justify-center mb-4">
+                  Import Project Data
+                </h2>
+                <CSVUploader />
+              </div>
+            </section>
+
+            {/* Schedule Configuration Section */}
+            <section className="card bg-base-100 shadow-lg">
+              <div className="card-body">
+                <h2 className="card-title justify-center mb-4">
+                  Schedule Configuration
+                </h2>
+
+                <div className="space-y-6">
+                  {/* Time Configuration */}
+                  <div className="space-y-4">
+                    <div className="form-control">
+                      <label className="label">
+                        <span className="label-text font-medium">
+                          Start Time
+                        </span>
+                      </label>
+                      <input
+                        type="datetime-local"
+                        value={startTime}
+                        onChange={(e) => setStartTime(e.target.value)}
+                        className="input input-bordered w-full"
+                      />
+                    </div>
+
+                    <div className="form-control">
+                      <label className="label">
+                        <span className="label-text font-medium">
+                          Duration per Project: {durationMinutes} minutes
+                        </span>
+                      </label>
+                      <input
+                        type="range"
+                        min="5"
+                        max="30"
+                        step="5"
+                        value={durationMinutes}
+                        onChange={(e) =>
+                          setDurationMinutes(Number(e.target.value))
+                        }
+                        className="range range-primary"
+                      />
+                    </div>
+                  </div>
+
+                  {/* Table Configuration */}
+                  <div className="form-control">
+                    <label className="label">
+                      <span className="label-text font-medium">
+                        Projects per Table: {projectsPerTable}
+                      </span>
+                    </label>
+                    <input
+                      type="range"
+                      min="1"
+                      max="20"
+                      value={projectsPerTable}
+                      onChange={(e) =>
+                        setProjectsPerTable(Number(e.target.value))
+                      }
+                      className="range range-primary"
+                    />
+                  </div>
+
+                  {/* Action Button */}
+                  <button
+                    onClick={() => {
+                      setStatus("loading");
+                      createTables.mutate({ projectsPerTable });
+                    }}
+                    disabled={status === "loading"}
+                    className={`btn w-full ${
+                      status === "idle"
+                        ? "btn-primary"
+                        : status === "loading"
+                        ? "btn-disabled loading"
+                        : status === "success"
+                        ? "btn-success"
+                        : "btn-error"
+                    }`}
+                    title={status === "error" ? errorMessage : ""}
+                  >
+                    {status === "loading"
+                      ? "Creating Schedule..."
+                      : status === "success"
+                      ? "Schedule Created!"
+                      : "Create Schedule"}
+                  </button>
+
+                  {/* Status Information */}
+                  {(status === "success" || status === "loading") && (
+                    <div className="mt-4 space-y-2 text-center">
+                      {judgingDuration && (
+                        <div className="stat-value text-lg">
+                          Total Duration: {judgingDuration}
+                        </div>
+                      )}
+                      {endTime && (
+                        <div className="stat-value text-lg">
+                          End Time: {endTime}
+                        </div>
+                      )}
+                      {numTables && (
+                        <div className="stat-value text-lg">
+                          Number of Tables: {numTables}
+                        </div>
+                      )}
+                    </div>
+                  )}
+                </div>
+              </div>
+            </section>
+          </div>
+        </main>
+      </Drawer>
+    </>
+  );
+};
+
+// Add server-side props function for access control
+export async function getServerSideProps(context: GetServerSidePropsContext) {
+  let output: GetServerSidePropsResult<Record<string, unknown>> = { props: {} };
+  output = rbac(
+    await getServerAuthSession(context),
+    [Role.ADMIN],
+    undefined,
+    output
+  );
+  return output;
+}
+
+export default JudgingPage;
